+++ tests/snapshot/mv2-basic-no-dynamic-import-rspack/background.js
@@ -95,7 +95,7 @@ __webpack_require__.d = function(exports, definition) {
           // return url for filenames not based on template
           
           // return url for filenames based on template
-          return "chunks-" + "8c99a9e37f1c38cc" + ".js";
+          return "chunks-" + "0819842744200c7a" + ".js";
         };
       
 })();
@@ -137,7 +137,7 @@ __webpack_require__.p = "";
 // webpack/runtime/rspack_version
 (() => {
 __webpack_require__.rv = function () {
-	return "1.1.3";
+	return "1.1.5";
 };
 
 })();
@@ -170,7 +170,7 @@ __webpack_require__.b = document.baseURI || self.location.href;
 })();
 // webpack/runtime/rspack_unique_id
 (() => {
-__webpack_require__.ruid = "bundler=rspack@1.1.3";
+__webpack_require__.ruid = "bundler=rspack@1.1.5";
 
 })();
 // webpack/runtime/chunk loader fallback