+++ tests/snapshot/mv2-basic-rspack/chunks-c29bd93b477d9a1c.js
@@ -139,7 +139,7 @@ __webpack_require__.p = "";
 // webpack/runtime/rspack_version
 (() => {
 __webpack_require__.rv = function () {
-	return "1.1.3";
+	return "1.1.5";
 };
 
 })();
@@ -168,7 +168,7 @@ var installedChunks = {"worker_js": 1,};
 })();
 // webpack/runtime/rspack_unique_id
 (() => {
-__webpack_require__.ruid = "bundler=rspack@1.1.3";
+__webpack_require__.ruid = "bundler=rspack@1.1.5";
 
 })();
 /************************************************************************/