+++ tests/snapshot/mv3-basic-no-dynamic-import-rspack/background.js
@@ -95,7 +95,7 @@ __webpack_require__.d = function(exports, definition) {
           // return url for filenames not based on template
           
           // return url for filenames based on template
-          return "chunks-" + "3930753a9cc30e23" + ".js";
+          return "chunks-" + "f2c2cf926f9a5dbb" + ".js";
         };
       
 })();
@@ -137,7 +137,7 @@ __webpack_require__.p = "";
 // webpack/runtime/rspack_version
 (() => {
 __webpack_require__.rv = function () {
-	return "1.1.3";
+	return "1.1.5";
 };
 
 })();
@@ -166,7 +166,7 @@ var installedChunks = {"background": 1,};
 })();
 // webpack/runtime/rspack_unique_id
 (() => {
-__webpack_require__.ruid = "bundler=rspack@1.1.3";
+__webpack_require__.ruid = "bundler=rspack@1.1.5";
 
 })();
 // webpack/runtime/chunk loader fallback