+++ tests/snapshot/mv3-basic-no-dynamic-import-rspack/chunks-f2c2cf926f9a5dbb.js
@@ -5,19 +5,6 @@ var __webpack_modules__ = ({
 module.exports = __webpack_require__.p + "6c5b191a31c5a9fc.txt";
 
 }),
-"./log.js": (function (__unused_webpack___webpack_module__, __webpack_exports__, __webpack_require__) {
-__webpack_require__.r(__webpack_exports__);
-__webpack_require__.d(__webpack_exports__, {
-  file: function() { return file; },
-  mod: function() { return /* reexport module object */ _log_js__WEBPACK_IMPORTED_MODULE_0__; }
-});
-/* ESM import */var _log_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./log.js */ "./log.js");
-
-
-const file = 'log.js'
-
-
-}),
 "./util.js": (function (__unused_webpack___webpack_module__, __webpack_exports__, __webpack_require__) {
 __webpack_require__.r(__webpack_exports__);
 __webpack_require__.d(__webpack_exports__, {
@@ -101,6 +88,18 @@ __webpack_require__.d = function(exports, definition) {
     }
 };
 })();
+// webpack/runtime/global
+(() => {
+__webpack_require__.g = (function () {
+	if (typeof globalThis === 'object') return globalThis;
+	try {
+		return this || new Function('return this')();
+	} catch (e) {
+		if (typeof window === 'object') return window;
+	}
+})();
+
+})();
 // webpack/runtime/has_own_property
 (() => {
 __webpack_require__.o = function (obj, prop) {
@@ -121,21 +120,32 @@ __webpack_require__.r = function(exports) {
 })();
 // webpack/runtime/public_path
 (() => {
-__webpack_require__.p = "/";
+__webpack_require__.p = "";
 
 })();
 // webpack/runtime/rspack_version
 (() => {
 __webpack_require__.rv = function () {
-	return "1.1.3";
+	return "1.1.5";
 };
 
 })();
 // webpack/runtime/publicPath
 (() => {
-if (__webpack_require__.webExtRt) {
-	__webpack_require__.p = __webpack_require__.webExtRt.runtime.getURL("/");
+let scriptUrl;
+if (__webpack_require__.g.importScripts) scriptUrl = __webpack_require__.g.location + "";
+const document = __webpack_require__.g.document;
+if (!scriptUrl && document?.currentScript) {
+	scriptUrl = document.currentScript.src;
 }
+// When supporting browsers where an automatic publicPath is not supported you must specify an output.publicPath manually via configuration
+// or pass an empty string ("") and set the __webpack_public_path__ variable from your code to use your own logic.
+if (!scriptUrl) {
+	if (__webpack_require__.webExtRt) scriptUrl = __webpack_require__.webExtRt.runtime.getURL("/");
+	else throw new Error("Automatic publicPath is not supported in this browser");
+}
+scriptUrl = scriptUrl.replace(/#.*$/, "").replace(/\?.*$/, "").replace(/\/[^\/]+$/, "/");
+__webpack_require__.p = scriptUrl;
 })();
 // webpack/runtime/import_scripts_chunk_loading
 (() => {
@@ -145,7 +155,7 @@ var installedChunks = {"worker_js": 1,};
 })();
 // webpack/runtime/rspack_unique_id
 (() => {
-__webpack_require__.ruid = "bundler=rspack@1.1.3";
+__webpack_require__.ruid = "bundler=rspack@1.1.5";
 
 })();
 /************************************************************************/
@@ -162,13 +172,6 @@ Promise.resolve()
     (0,_util_js__WEBPACK_IMPORTED_MODULE_0__.log)('Worker Test A: import.meta.url', () => {
       const url = new URL(/* asset import */__webpack_require__(/*! ./test.txt */ "./test.txt"), __webpack_require__.b).toString()
       ;(0,_util_js__WEBPACK_IMPORTED_MODULE_0__.test)(url.includes('-extension://'), "new URL('./test.txt', import.meta.url)\n", url)
-    })
-  )
-  .then(
-    (0,_util_js__WEBPACK_IMPORTED_MODULE_0__.log)('Worker Test B: dynamic import', async () => {
-      console.log("await import('./log.js')\n")
-      const mod = await Promise.resolve(/*! import() eager */).then(__webpack_require__.bind(__webpack_require__, /*! ./log.js */ "./log.js"))
-      ;(0,_util_js__WEBPACK_IMPORTED_MODULE_0__.test)('file' in mod, mod)
     })
   )
   .then(() => new Promise((resolve) => setTimeout(resolve, 100)))