+++ tests/snapshot/mv3-basic-public-path-rspack/background.js
@@ -108,7 +108,7 @@ __webpack_require__.d = function(exports, definition) {
           // return url for filenames not based on template
           
           // return url for filenames based on template
-          return "chunks-" + "3b6fe31d3a6fa640" + ".js";
+          return "chunks-" + "576cccd3cc817224" + ".js";
         };
       
 })();
@@ -138,7 +138,7 @@ __webpack_require__.p = "/";
 // webpack/runtime/rspack_version
 (() => {
 __webpack_require__.rv = function () {
-	return "1.1.3";
+	return "1.1.5";
 };
 
 })();
@@ -156,7 +156,7 @@ var installedChunks = {"background": 1,};
 })();
 // webpack/runtime/rspack_unique_id
 (() => {
-__webpack_require__.ruid = "bundler=rspack@1.1.3";
+__webpack_require__.ruid = "bundler=rspack@1.1.5";
 
 })();
 // webpack/runtime/chunk loader fallback