+++ tests/snapshot/mv3-basic-public-path-rspack/content.js
@@ -127,7 +127,7 @@ __webpack_require__.p = "/";
 // webpack/runtime/rspack_version
 (() => {
 __webpack_require__.rv = function () {
-	return "1.1.3";
+	return "1.1.5";
 };
 
 })();
@@ -149,7 +149,7 @@ __webpack_require__.b = document.baseURI || self.location.href;
 })();
 // webpack/runtime/rspack_unique_id
 (() => {
-__webpack_require__.ruid = "bundler=rspack@1.1.3";
+__webpack_require__.ruid = "bundler=rspack@1.1.5";
 
 })();
 /************************************************************************/