+++ tests/snapshot/mv3-basic-weak-runtime-rspack/background.js
@@ -111,7 +111,7 @@ __webpack_require__.d = function(exports, definition) {
           // return url for filenames not based on template
           
           // return url for filenames based on template
-          return "chunks-" + "21226f49a8a3b562" + ".js";
+          return "chunks-" + "08f25013a1db988c" + ".js";
         };
       
 })();
@@ -141,7 +141,7 @@ __webpack_require__.p = "/";
 // webpack/runtime/rspack_version
 (() => {
 __webpack_require__.rv = function () {
-	return "1.1.3";
+	return "1.1.5";
 };
 
 })();
@@ -159,7 +159,7 @@ var installedChunks = {"background": 1,};
 })();
 // webpack/runtime/rspack_unique_id
 (() => {
-__webpack_require__.ruid = "bundler=rspack@1.1.3";
+__webpack_require__.ruid = "bundler=rspack@1.1.5";
 
 })();
 // webpack/runtime/chunk loader fallback