+++ tests/snapshot/mv3-hmr-rspack/background.js
@@ -189,7 +189,7 @@ __webpack_require__.d = function(exports, definition) {
           // return url for filenames not based on template
           
           // return url for filenames based on template
-          return "chunks-" + "5594f47c3921be86" + ".js";
+          return "chunks-" + "582a1acde5595e2f" + ".js";
         };
       
 })();
@@ -203,7 +203,7 @@ __webpack_require__.hu = function (chunkId) {
 // webpack/runtime/get_full_hash
 (() => {
 __webpack_require__.h = function () {
-	return "abb9ef3d4e15229e";
+	return "c6048ed87d46cc4b";
 };
 
 })();
@@ -618,7 +618,7 @@ __webpack_require__.p = "";
 // webpack/runtime/rspack_version
 (() => {
 __webpack_require__.rv = function () {
-	return "1.1.3";
+	return "1.1.5";
 };
 
 })();
@@ -1106,7 +1106,7 @@ __webpack_require__.hmrM = function () {
 })();
 // webpack/runtime/rspack_unique_id
 (() => {
-__webpack_require__.ruid = "bundler=rspack@1.1.3";
+__webpack_require__.ruid = "bundler=rspack@1.1.5";
 
 })();
 // webpack/runtime/chunk loader fallback